backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
psycopg2-binary==2.9.10
python-dotenv==1.0.1
anthropic==0.40.0
sentence-transformers==3.3.1
slowapi==0.5.7
